# Enable cat for nvfuser >= 0.1.7