--- a/thunder/tests/test_nvfuser.py
+++ b/thunder/tests/test_nvfuser.py
@@ -356,11 +356,11 @@ def test_cse_rematerialization(executor, device, _):
     assert len(fusion_bsyms) == 11
     # fusion groups 1 and 6 correspond with the apply_rotary_emb function
     # Nvfuser with recomputation should use precomputed cos and sin values.
-    assert len(fusion_bsyms[1].args) == len(fusion_bsyms[7].args)
+    assert len(fusion_bsyms[1].args) == len(fusion_bsyms[6].args)
     assert fusion_bsyms[1].subsymbols[0].output.name == "freqs_cos"
     assert fusion_bsyms[1].subsymbols[1].output.name == "freqs_sin"
-    assert fusion_bsyms[7].subsymbols[0].output.name == "freqs_cos"
-    assert fusion_bsyms[7].subsymbols[1].output.name == "freqs_sin"
+    assert fusion_bsyms[6].subsymbols[0].output.name == "freqs_cos"
+    assert fusion_bsyms[6].subsymbols[1].output.name == "freqs_sin"
 
 
 # Tests that two separated nvFuser regions can be merged when they don't depend